Remove unused Router interface Rename RequestRouter to Router which implements the Handler interface
