+++ src/main/java/io/github/christophermanahan/carnitas/Main.java
@@ -29,7 +29,7 @@ public class Main {
     }
 
     private static Handler router() {
-        return new RequestRouter()
+        return new Router()
           .get("/simple_get", (HTTPRequest request) -> new HTTPResponse(HTTPResponse.StatusCode.OK))
           .head("/simple_get", (HTTPRequest request) -> new HTTPResponse(HTTPResponse.StatusCode.OK))
           .post("/simple_post", (HTTPRequest request) -> new HTTPResponse(HTTPResponse.StatusCode.CREATED)
