+++ src/main/java/io/github/christophermanahan/carnitas/Router.java
@@ -1,13 +1,56 @@
 package io.github.christophermanahan.carnitas;
 
+import java.util.HashMap;
 import java.util.function.Function;
 
-interface Router {
-    RequestRouter get(String uri, Function<HTTPRequest, HTTPResponse> handler);
+class Router implements Handler {
+    private final HashMap<Route, Function<HTTPRequest, HTTPResponse>> map;
+    static final String GET = "GET";
+    static final String HEAD = "HEAD";
+    static final String POST = "POST";
 
-    RequestRouter head(String uri, Function<HTTPRequest, HTTPResponse> handler);
+    Router() {
+        this.map = new HashMap<>();
+    }
 
-    RequestRouter post(String uri, Function<HTTPRequest, HTTPResponse> handler);
+    Router get(String uri, Function<HTTPRequest, HTTPResponse> handler) {
+        map.put(new Route(GET, uri), handler);
+        return this;
+    }
 
-    HTTPResponse handle(HTTPRequest request);
+    Router head(String uri, Function<HTTPRequest, HTTPResponse> handler) {
+        map.put(new Route(HEAD, uri), handler);
+        return this;
+    }
+
+    Router post(String uri, Function<HTTPRequest, HTTPResponse> handler) {
+        map.put(new Route(POST, uri), handler);
+        return this;
+    }
+
+    public HTTPResponse handle(HTTPRequest request) {
+        if (routeAdded(request)) {
+            return handleRoute(request);
+        } else {
+            return new HTTPResponse(HTTPResponse.StatusCode.NOT_FOUND);
+        }
+    }
+
+    private HTTPResponse handleRoute(HTTPRequest request) {
+        return map.get(
+          map.keySet().stream()
+            .filter(route -> route.equals(route(request)))
+            .findFirst()
+            .get()
+        ).apply(request);
+    }
+
+    private boolean routeAdded(HTTPRequest request) {
+        return map.keySet().stream()
+          .anyMatch(route -> route.equals(route(request)));
+    }
+
+    private Route route(HTTPRequest request) {
+        return new Route(request.method(), request.uri());
+    }
 }
